Move main logic to post logic
Since gathering data need to be after all jobs and steps complete, it
must need to be execute under post

haha

a

a

a
