+++ action.yml
@@ -20,3 +20,4 @@ outputs:
 runs:
   using: node20
   main: dist/index.js
+  post: dist/post.js
