+++ rollup.config.ts
@@ -5,17 +5,17 @@ import nodeResolve from '@rollup/plugin-node-resolve'
 import typescript from '@rollup/plugin-typescript'
 import json from '@rollup/plugin-json'
 
-const config = {
-  input: 'src/index.ts',
-  output: {
-    esModule: true,
-    // file: 'dist/index.js',
-    dir: 'dist',
-    format: 'es',
-    sourcemap: true
-  },
-  plugins: [nodeResolve(), commonjs(), json(), typescript()]
-  // external: ['@mermaid-js/mermaid-cli']
-}
+const config = [
+  {
+    input: ['src/index.ts', 'src/post.ts'],
+    output: {
+      esModule: true,
+      dir: 'dist',
+      format: 'es',
+      sourcemap: true
+    },
+    plugins: [nodeResolve(), commonjs(), json(), typescript()]
+  }
+]
 
 export default config
